+++ examples/linguistic_features/named_entity_recognition.rb
@@ -10,7 +10,7 @@ headings = ["text", "start", "end", "label"]
 rows = []
 
 doc.ents.each do |ent|
-  rows << [ent.text, ent.start_char, ent.end_char, ent.label_]
+  rows << [ent.text, ent.start_char, ent.end_char, ent.label]
 end
 
 table = Terminal::Table.new rows: rows, headings: headings