+++ examples/linguistic_features/retokenize_1.rb
@@ -12,7 +12,7 @@ rows = []
 doc.retokenize(doc[4].left_edge.i, doc[4].right_edge.i)
 
 doc.each do |token|
-  rows << [token.text, token.pos_, token.dep_, token.head.text]
+  rows << [token.text, token.pos, token.dep, token.head.text]
 end
 
 table = Terminal::Table.new rows: rows, headings: headings