+++ examples/linguistic_features/retokenize_2.rb
@@ -6,11 +6,11 @@ nlp = Spacy::Language.new("en_core_web_sm")
 sentence = "I live in New York"
 doc = nlp.read(sentence)
 
-puts "Before: " + doc.tokens.collect{|t| t}.join(", ")
+puts "Before: " + doc.tokens.map(&:text).join(", ")
 
 doc.retokenize(3, 4)
 
-puts "After: " + doc.tokens.collect{|t| t}.join(", ")
+puts "After: " + doc.tokens.map(&:text).join(", ")
 
 # Before: I, live, in, New, York
 # After: I, live, in, New York