+++ examples/linguistic_features/rule_based_morphology.rb
@@ -6,7 +6,7 @@ nlp = Spacy::Language.new("en_core_web_sm")
 doc = nlp.read("Where are you?")
 
 puts "Morph features of the third word: " + doc[2].morph.to_s
-puts "POS of the third word: " + doc[2].pos_.to_s
+puts "POS of the third word: " + doc[2].pos
 
 # Morph features of the third word: Case=Nom|Person=2|PronType=Prs
 # POS of the third word: PRON